+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All updates to the mods will be noted in this file
 ---
 
 ## [Unreleased]
+- Added "rainbow" trail
+- Added menu toggle to switch from text to rainbow
+- Added a popup bubble when walking next to npcs
+- Improved trail edit menu to be formatted nicer
+- removed custom furniture as well as JSON assets, no longer supported
+- added npc customization (currently linus as JFS, going to be wizard)
+- added basic star trail
 
 ## [1.0.0] - 2025-07-16
 - First commit created